Make the type repo match and fix a mistake in brewery repo

+++ src/main/java/edu/rlperez/java_general_web_demo/repositories/BreweryTypeRepository.java
@@ -3,5 +3,5 @@ package edu.rlperez.java_general_web_demo.repositories;
 import edu.rlperez.java_general_web_demo.models.BreweryType;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BreweryTypeRepository extends CrudRepository<BreweryType, Long> {
+public interface BreweryTypeRepository extends CrudRepository<BreweryType, Integer> {
 }
